backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
decorator==5.1.1
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
django-environ==0.11.2
django-filter==23.3
django-modeltranslation==0.18.11
django-parler==2.3
django-parler-rest==2.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
docopt==0.6.2
drf-yasg==1.21.7
environs==9.5.0
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.0
idna==3.4
inflection==0.5.1
Js2Py==0.74
jsonschema-specifications==2023.7.1
marshmallow==3.20.1
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pandas==2.2.2
Pillow==10.1.0
pipwin==0.5.2
psycopg2==2.9.9
pycparser==2.21
pyjsparser==2.7.1
PyJWT==2.8.0
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
ratelim==0.1.6
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.6
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
soupsieve==2.5
spatialite==0.0.3
sqlparse==0.4.4
stripe==10.5.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.0.7
